[PASpeakerDetector] Change init to receive full pipeline model
The segmentation model, embeddings model, clustering method and pipeline
params meters of a pyannote diarization pipeline shouldn't really be
customized independently of each other, they have to be trained and
optimized jointly.

Therefore it makes more sense to init PASpeakerDetector detector with
a full pipeline config (which contains the segmentation and embeddings
models, the clustering method and the hyperparameters).