CI: Do not treat warnings as errors in RTD
Since the last release, Sphinx produces some warnings of the likes:

`WARNING: Inline interpreted text or phrase reference start-string without end-string.`

Since the content of the affected docstrings has not changed,
this is likely a regression in Sphinx or one of its plugins.

In the intrest of moving forward, just disable `fail_on_warning`
to allow for the next release to happen.